refactor(abilities): split use_ability/cone_aoe along flow phases + cone seams
Tech-debt #529: bring two oversized cell-ability files under the 700-line
hard cap by pure code movement along natural seams. Zero public-surface
change — the surviving `mod.rs` files re-export every public item so all
`crate::cell::abilities::…` paths stay identical.

use_ability/ (was 934-line mod.rs):
- handle.rs        — the main `handle_use_ability` validate→consume→fire→resolve flow
- auto_reload.rs   — `maybe_trigger_auto_reload` (+ test re-export)
- kill_credit.rs   — `handle_use_ability_with_kill_credit`
- weapon_redirect.rs — read-only archetype-default→active-weapon RANGED redirect
- mod.rs           — declarations + re-exports

cone_aoe.rs (845 lines) → cone_aoe/:
- geometry.rs       — `is_cone_effect` + `collect_cone_targets`
- flag_categories.rs — `log_effect_flag_categories`
- fan_out.rs        — `fan_out_cone_effects`
- tests.rs          — moved test module (import paths fixed only)
- mod.rs            — declarations + re-exports

The archetype-weapon-redirect block was extracted from `handle_use_ability`
as the pure read-only `resolve_weapon_redirect` helper (no borrow reorder,
no behavior change) — this brings handle.rs under the hard cap while keeping
the deliberately-monolithic core fire function intact. Added net-new unit
tests for the extracted helper (pass-through, redirect, no-weapon, same-id
no-op). No existing test logic changed.

Behavior-preserving · re-export discipline · zero public-surface change.

Part of #529.
